c.d

+++ SoundEncoderDecoder/Modulation/IModulator.cs
@@ -4,8 +4,8 @@ using System.IO;
 namespace SoundEncoderDecoder.Modulation {
     public interface IModulator {
         public SampleRateType SampleRate { get; }
-        public float CarrierFrequency { get; }
-        public float BitDuration { get; }
+        public double CarrierFrequency { get; }
+        public double BitDuration { get; }
         void WriteZero(MemoryStream ms);
         void WriteOne(MemoryStream ms);
     }
